[Impeller] make VerticesGeometry delegate to the DL class
