+++ impeller/aiks/canvas.cc
@@ -18,7 +18,6 @@
 #include "impeller/entity/contents/vertices_contents.h"
 #include "impeller/entity/geometry.h"
 #include "impeller/geometry/path_builder.h"
-#include "impeller/geometry/vertices.h"
 
 namespace impeller {
 
@@ -373,11 +372,9 @@ void Canvas::DrawTextFrame(const TextFrame& text_frame,
   GetCurrentPass().AddEntity(entity);
 }
 
-void Canvas::DrawVertices(const Vertices& vertices,
+void Canvas::DrawVertices(std::unique_ptr<VerticesGeometry> vertices,
                           BlendMode blend_mode,
                           Paint paint) {
-  auto geometry = Geometry::MakeVertices(vertices);
-
   Entity entity;
   entity.SetTransformation(GetCurrentTransformation());
   entity.SetStencilDepth(GetStencilDepth());
@@ -386,7 +383,7 @@ void Canvas::DrawVertices(const Vertices& vertices,
   if (paint.color_source.has_value()) {
     auto& source = paint.color_source.value();
     auto contents = source();
-    contents->SetGeometry(std::move(geometry));
+    contents->SetGeometry(std::move(vertices));
     contents->SetAlpha(paint.color.alpha);
     entity.SetContents(paint.WithFilters(std::move(contents), true));
   } else {
@@ -394,7 +391,7 @@ void Canvas::DrawVertices(const Vertices& vertices,
         std::make_shared<VerticesContents>();
     contents->SetColor(paint.color);
     contents->SetBlendMode(blend_mode);
-    contents->SetGeometry(std::move(geometry));
+    contents->SetGeometry(std::move(vertices));
     entity.SetContents(paint.WithFilters(std::move(contents), true));
   }
