+++ impeller/entity/entity_unittests.cc
@@ -1203,77 +1203,6 @@ TEST_P(EntityTest, BorderMaskBlurCoverageIsCorrect) {
   }
 }
 
-TEST_P(EntityTest, DrawVerticesSolidColorTrianglesWithoutIndices) {
-  std::vector<Point> positions = {Point(100, 300), Point(200, 100),
-                                  Point(300, 300)};
-  std::vector<Color> colors = {Color::White(), Color::Green(), Color::White()};
-
-  Vertices vertices = Vertices(positions, {} /* indices */, colors,
-                               VertexMode::kTriangle, Rect(100, 100, 300, 300));
-
-  std::shared_ptr<VerticesContents> contents =
-      std::make_shared<VerticesContents>();
-  contents->SetGeometry(Geometry::MakeVertices(vertices));
-  contents->SetBlendMode(BlendMode::kSourceOver);
-  contents->SetColor(Color::Red().WithAlpha(0.5));
-
-  Entity e;
-  e.SetTransformation(Matrix::MakeScale(GetContentScale()));
-  e.SetContents(contents);
-
-  ASSERT_TRUE(OpenPlaygroundHere(e));
-}
-
-TEST_P(EntityTest, DrawVerticesLinearGradientWithoutIndices) {
-  std::vector<Point> positions = {Point(100, 300), Point(200, 100),
-                                  Point(300, 300)};
-
-  Vertices vertices = Vertices(positions, {} /* indices */, {} /* colors */,
-                               VertexMode::kTriangle, Rect(100, 100, 300, 300));
-
-  std::vector<Color> colors = {Color{0.9568, 0.2627, 0.2118, 1.0},
-                               Color{0.1294, 0.5882, 0.9529, 1.0}};
-  std::vector<Scalar> stops = {0.0, 1.0};
-  Matrix matrix = {
-      1, 0, 0, 0,  //
-      0, 1, 0, 0,  //
-      0, 0, 1, 0,  //
-      0, 0, 0, 1   //
-  };
-  auto contents = std::make_shared<LinearGradientContents>();
-  contents->SetEndPoints({100, 100}, {300, 300});
-  contents->SetColors(std::move(colors));
-  contents->SetStops(std::move(stops));
-  contents->SetTileMode(Entity::TileMode::kRepeat);
-  contents->SetMatrix(matrix);
-  contents->SetGeometry(Geometry::MakeVertices(vertices));
-
-  Entity e;
-  e.SetTransformation(Matrix::MakeScale(GetContentScale()));
-  e.SetContents(contents);
-
-  ASSERT_TRUE(OpenPlaygroundHere(e));
-}
-
-TEST_P(EntityTest, DrawVerticesSolidColorTrianglesWithIndices) {
-  std::vector<Point> positions = {Point(100, 300), Point(200, 100),
-                                  Point(300, 300), Point(200, 500)};
-  std::vector<uint16_t> indices = {0, 1, 2, 0, 2, 3};
-
-  Vertices vertices = Vertices(positions, indices, {} /* colors */,
-                               VertexMode::kTriangle, Rect(100, 100, 300, 300));
-
-  std::shared_ptr<VerticesContents> contents =
-      std::make_shared<VerticesContents>();
-  contents->SetGeometry(Geometry::MakeVertices(vertices));
-  contents->SetColor(Color::White());
-  Entity e;
-  e.SetTransformation(Matrix::MakeScale(GetContentScale()));
-  e.SetContents(contents);
-
-  ASSERT_TRUE(OpenPlaygroundHere(e));
-}
-
 TEST_P(EntityTest, DrawAtlasNoColor) {
   // Draws the image as four squares stiched together.
   auto atlas = CreateTextureForFixture("bay_bridge.jpg");
