+++ impeller/geometry/BUILD.gn
@@ -39,8 +39,6 @@ impeller_component("geometry") {
     "type_traits.h",
     "vector.cc",
     "vector.h",
-    "vertices.cc",
-    "vertices.h",
   ]
 
   deps = [ "//flutter/fml" ]
